requirements.txt
crewai>=0.1.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
requests>=2.31.0
PyMuPDF>=1.23.0
crawl4ai>=0.3.0
flask>=2.3.0
flask-socketio>=5.3.0
arxiv>=2.1.0
psycopg2-binary>=2.9.0
sentence-transformers>=2.2.0
langchain>=0.1.0
langchain-community>=0.0.10
